Share the EMM386 page-table selector
Store only the varying page-table offset for each of the 52 physical windows. Mapping code now loads the common PAGET_GSEL selector explicitly instead of retaining an identical selector word in every table entry.

This removes 104 resident data bytes for four code bytes and gains another four through alignment. Focused frame, banking, sparse-page, EMS 4.0, maximum-option, UMB, rollback, runtime-mode, and warm-reboot tests pass. The paired VC capture measures EMM386 at 13,456 bytes and the largest conventional block at 594,848 bytes.

+++ src/MEMM/EMM/EMM40.C
@@ -95,7 +95,7 @@ extern char cntxt_pages;		/* pages in context */
 extern char cntxt_bytes;		/* bytes in context */
 extern unsigned short PF_Base;
 /*extern unsigned short VM1_EMM_Offset;*/
-extern long	page_frame_base[];
+extern unsigned short page_frame_base[];
 extern char	EMM_MPindex[];
 extern char	OSEnabled;			/* OS/E function flag */
 extern long	OSKey;				/* Key for OS/E function */
